+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.2.7
 
 - Added constructor with custom client for `GSheets`
-- Added `copyTo` method to `GSheets`
+- Added `close` method to `GSheets`
 
 ## 0.2.6